third_party/WebKit/LayoutTests/inspector/sources/debugger-frameworks/frameworks-jquery-expected.txt


Tests framework blackboxing feature on jQuery.

[page] onTestEvent1
[page] onTestEvent2
Set timer for test function.
Call stack:
    0) testFunction (frameworks-jquery.html:15)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepInto...
Executing StepInto...
Call stack:
    0) onEachScript (frameworks-jquery.html:33)
  * 1) each (jquery-1.11.1.min.js:2)
  * 2) each (jquery-1.11.1.min.js:2)
    3) testFunction (frameworks-jquery.html:17)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepInto...
Executing StepInto...
Call stack:
    0) onEachScript (frameworks-jquery.html:33)
  * 1) each (jquery-1.11.1.min.js:2)
  * 2) each (jquery-1.11.1.min.js:2)
    3) testFunction (frameworks-jquery.html:17)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepOver...
Executing StepOver...
Call stack:
    0) onEachScript (frameworks-jquery.html:33)
  * 1) each (jquery-1.11.1.min.js:2)
  * 2) each (jquery-1.11.1.min.js:2)
    3) testFunction (frameworks-jquery.html:17)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepOver...
Executing StepOver...
Call stack:
    0) onEachScript (frameworks-jquery.html:33)
  * 1) each (jquery-1.11.1.min.js:2)
  * 2) each (jquery-1.11.1.min.js:2)
    3) testFunction (frameworks-jquery.html:17)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepOut...
Call stack:
    0) testFunction (frameworks-jquery.html:18)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepInto...
Call stack:
    0) onTestEvent1 (frameworks-jquery.html:23)
  * 1) dispatch (jquery-1.11.1.min.js:3)
  * 2) r.handle (jquery-1.11.1.min.js:3)
  * 3) trigger (jquery-1.11.1.min.js:3)
  * 4)  (jquery-1.11.1.min.js:3)
  * 5) each (jquery-1.11.1.min.js:2)
  * 6) each (jquery-1.11.1.min.js:2)
  * 7) trigger (jquery-1.11.1.min.js:3)
    8) testFunction (frameworks-jquery.html:18)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepOver...
Executing StepOver...
Call stack:
    0) onTestEvent2 (frameworks-jquery.html:28)
  * 1) dispatch (jquery-1.11.1.min.js:3)
  * 2) r.handle (jquery-1.11.1.min.js:3)
  * 3) trigger (jquery-1.11.1.min.js:3)
  * 4)  (jquery-1.11.1.min.js:3)
  * 5) each (jquery-1.11.1.min.js:2)
  * 6) each (jquery-1.11.1.min.js:2)
  * 7) trigger (jquery-1.11.1.min.js:3)
    8) testFunction (frameworks-jquery.html:18)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>
